add cast

+++ data_iterator_paddle.py
@@ -39,7 +39,7 @@ def reader_data(source0, batch_size0=256, max_batch_size0=20):
                 ss[264]=int(float(ss[264])*1000000)
                 source.append(ss[:-1])
                 target.append(ss[-1])
-            source =  np.array(source, np.int32)
+            source =  np.array(source, np.float32).astype("int32")
             target =  np.array(target, np.float32)
             yield source, target
